+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_coop_per_player_loadout/gwo_loadouts.js
@@ -196,7 +196,6 @@ function gwoLoadouts() {
       }
     );
   } catch (e) {
-    console.error(e);
     console.error(
       "Galactic War Overhaul (GWO): " + (e.stack || e.message || e)
     );