+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_play/gwo_panel.js
@@ -370,7 +370,6 @@ function gwoWarInfoPanel(gwoSettings) {
       }
     );
   } catch (e) {
-    console.error(e);
     console.error(
       "Galactic War Overhaul (GWO): " + (e.stack || e.message || e)
     );