+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_play/referee.js
@@ -105,7 +105,6 @@ function gwoRefereeChanges() {
       }
     );
   } catch (e) {
-    console.error(e);
     console.error(
       "Galactic War Overhaul (GWO): " + (e.stack || e.message || e)
     );