+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_start/ui.js
@@ -446,7 +446,6 @@ function gwoUI() {
       return model.mode() || loc("!LOC:Galactic War Overhaul");
     });
   } catch (e) {
-    console.error(e);
     console.error(
       "Galactic War Overhaul (GWO): " + (e.stack || e.message || e)
     );