+++ ui/mods/com.pa.quitch.gwaioverhaul/live_game/menu.js
@@ -103,7 +103,6 @@ function gwoLiveGameMenu() {
       });
     });
   } catch (e) {
-    console.error(e);
     console.error(
       "Galactic War Overhaul (GWO): " + (e.stack || e.message || e)
     );