+++ ui/mods/com.pa.quitch.gwaioverhaul/live_game/win_conditions_text.js
@@ -1,0 +1,35 @@
+define(function () {
+  // options is the game_options object the server echoes to live_game clients;
+  // loc is passed in so the module stays loadable under the Node test harness
+  return function (options, loc) {
+    if (!options || options.game_type !== "Galactic War") {
+      return "";
+    }
+
+    var modifiers = [];
+    // the server's sudden death check ignores eradication_mode
+    if (options.sudden_death_mode) {
+      modifiers.push(loc("!LOC:Sudden Death"));
+    } else if (options.eradication_mode) {
+      var targets = [loc("!LOC:Commander")];
+      if (options.eradication_mode_sub_commanders) {
+        targets.push(loc("!LOC:Colonel"));
+      }
+      if (options.eradication_mode_factories) {
+        targets.push(loc("!LOC:Factory"));
+      }
+      if (options.eradication_mode_fabricators) {
+        targets.push(loc("!LOC:Fabber"));
+      }
+      modifiers.push(loc("!LOC:Eradicate") + ": " + targets.join(", "));
+    }
+    if (options.bounty_mode) {
+      var bounty = loc("!LOC:Bounties");
+      if (typeof options.bounty_value === "number") {
+        bounty += " x" + options.bounty_value;
+      }
+      modifiers.push(bounty);
+    }
+    return modifiers.join(" | ");
+  };
+});