+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # CHANGELOG
 
+## v6.12.0 - 2026-08-24
+
+### Added
+
+- Battles now display their win-condition modifiers
+
+### Changed
+
+- Eradication mode will show Commander in its list as a reminder
+
+### Bugfix
+
+- Don't show Eradication mode in the panel if Sudden Death is active as the latter overrides the former
+
 ## v6.11.0 - 2026-08-23
 
 ### Added